introduces 'AlleleCounts'
The type 'AlleleCounts' is a pair of two counters in the range [0..2].
It has an instance of 'Storable' that puts both counters into a single
'Word8'.  Exchanging the 'Unboxed.Vector Word8' in 'Variant' to a
'Storable.Vector AlleleCounts' means that the same compact encoding as
before is used, but the bit twiddling is localized to a single place and
all other code becomes much clearer and more type safe.